Config update (#50)
* remote p&p

* add bfeapi.js functionality

* disabling buttons

* fix lint errors

* fix lint errors
