+++ package.json
@@ -6,7 +6,7 @@
     "editor",
     "rdf"
   ],
-  "version": "0.4.0",
+  "version": "1.0.0",
   "engines": {
     "node": ">=10.0.0"
   },
