+++ static/js/config-dev.js
@@ -18,190 +18,6 @@ function myCB(data) {
   document.body.scrollTop = document.documentElement.scrollTop = 0;
 }
 
-function save(data, bfelog, callback){
-  var $messagediv = $('<div>', {id: "bfeditor-messagediv", class:"col-md-10 main"});
-
-  var url = config.url + "/verso/api/bfs/upsertWithWhere?where=%7B%22name%22%3A%20%22"+data.name+"%22%7D";
-
-  $.ajax({
-    url: url,
-    type: "POST",
-    data:JSON.stringify(data),
-    dataType: "json",
-    contentType: "application/json; charset=utf-8"
-  }).done(function (data) {
-    document.body.scrollTop = document.documentElement.scrollTop = 0;
-    bfelog.addMsg(new Error(), "INFO", "Saved " + data.id);
-    var $messagediv = $('<div>', {id: "bfeditor-messagediv"});
-    var decimaltranslator = window.ShortUUID("0123456789");
-    var resourceName = "e" + decimaltranslator.fromUUID(data.name);
-    $messagediv.append('<div class="alert alert-info"><strong>Description saved:</strong><a href='+data.url+'>'+resourceName+'</a></div>')
-    $('#bfeditor-formdiv').empty();
-    $('#save-btn').remove();
-    $messagediv.insertBefore('.nav-tabs');
-    $('#bfeditor-previewPanel').remove();
-    $('.nav-tabs a[href="#browse"]').tab('show')
-    bfeditor.bfestore.store = [];
-    window.location.hash = "";
-    callback(true, data.name);
-  }).fail(function (XMLHttpRequest, textStatus, errorThrown){
-    bfelog.addMsg(new Error(), "ERROR", "FAILED to save");
-    bfelog.addMsg(new Error(), "ERROR", "Request status: " + textStatus + "; Error msg: " + errorThrown);
-    $messagediv.append('<div class="alert alert-danger"><strong>Save Failed:</strong>'+errorThrown+'</span>');
-    $messagediv.insertBefore('.nav-tabs');
-  }).always(function(){                       
-    $('#table_id').DataTable().ajax.reload();
-  });
-}
-
-function publish(data, rdfxml, savename, bfelog, callback){
-  var $messagediv = $('<div>', {id: "bfeditor-messagediv", class:"col-md-10 main"});
-
-  //var url = "http://mlvlp04.loc.gov:8201/bibrecs/bfe2mets.xqy";
-  var url = config.url + "/profile-edit/server/publish";
-  var saveurl = "/verso/api/bfs/upsertWithWhere?where=%7B%22name%22%3A%20%22"+savename+"%22%7D";
-
-  var savedata = {};
-  savedata.name = savename;
-  savedata.profile = data.profile;
-  savedata.url = data.url;
-  savedata.created = data.created;
-  savedata.modified = data.modified;
-  savedata.status = data.status;
-  savedata.rdf = data.rdf;
-
-  data.rdfxml = JSON.stringify(rdfxml);
-  
-  $.when(
-    $.ajax({
-      url: saveurl,
-      type: "POST",
-      data:JSON.stringify(savedata),
-      dataType: "json",
-      contentType: "application/json; charset=utf-8"
-    }),
-    $.ajax({
-      url: url,
-      type: "POST",
-      data: JSON.stringify(data),
-      dataType: "json",
-      contentType: "application/json; charset=utf-8"
-    })).done(function (savedata, publishdata) {
-      document.body.scrollTop = document.documentElement.scrollTop = 0;
-      bfelog.addMsg(new Error(), "INFO", "Published " + publishdata[0].name);
-      var $messagediv = $('<div>', {id: "bfeditor-messagediv"});
-      $messagediv.append('<div class="alert alert-info"><strong>Description submitted for posting:</strong><a href=' + config.basedbURI + "/" + publishdata[0].objid+'>'+publishdata[0].lccn+'</a></div>');
-      $('#bfeditor-formdiv').empty();
-      $('#save-btn').remove();
-      $messagediv.insertBefore('.nav-tabs');
-      $('#bfeditor-previewPanel').remove();
-      $('.nav-tabs a[href="#browse"]').tab('show')
-      bfeditor.bfestore.store = [];
-      window.location.hash = "";
-      callback(true, data.name);                
-    }).fail(function (XMLHttpRequest, textStatus, errorThrown){
-      bfelog.addMsg(new Error(), "ERROR", "FAILED to save");
-      bfelog.addMsg(new Error(), "ERROR", "Request status: " + textStatus + "; Error msg: " + errorThrown);
-      $messagediv.append('<div class="alert alert-danger"><strong>Save Failed:</strong>'+errorThrown+'</span>');
-      $messagediv.insertBefore('#bfeditor-previewPanel');
-    }).always(function(){
-      $('#table_id').DataTable().ajax.reload();
-    });
-}
-
-
-function retrieve(uri, bfestore, loadtemplates, bfelog, callback){
-  var $messagediv = $('<div>', {id: "bfeditor-messagediv", class:"col-md-10 main"});
-  var url = config.url + "/profile-edit/server/whichrt";
-  var dType = (bfestore.state == 'loadmarc') ? 'xml' : 'json';
-
-  $.ajax({
-    dataType: dType,
-    type: "GET",
-    async: false,
-    data: { uri: uri},
-    url: url,
-    success: function (data) {
-      bfelog.addMsg(new Error(), "INFO", "Fetched external source baseURI " + uri);
-      bfelog.addMsg(new Error(), "DEBUG", "Source data", data);
-      
-      if (dType == 'xml') {
-        var recCount = $('zs\\:numberOfRecords', data).text();
-        if (recCount != '0') {
-          var rdfrec  = $('zs\\:recordData', data).html();
-          var recid = $('bf\\:Local > rdf\\:value', data).html()
-          recid = recid.padStart(9, '0');
-          bfestore.rdfxml2store(rdfrec, loadtemplates, recid, callback);
-        } else {
-          var q = uri.replace(/.+query=(.+?)&.+/, "$1");
-          $nohits = $('<div class="modal" tabindex="-1" role="dialog" id="nohits"><div class="modal-dialog" role="document"><div class="modal-content"> \
-          <div class="modal-header">No Record Found!</div><div class="modal-body"><p>Query: "' + q + '"</p></div> \
-          <div class="modal-footer"><button type="button" class="btn btn-primary" data-dismiss="modal">OK</button></div></div></div></div>');
-          $nohits.modal('show');
-        }
-      } else {
-        bfestore.store = bfestore.jsonldcompacted2store(data, function(expanded) {
-          bfestore.store = [];
-          tempstore = bfestore.jsonld2store(expanded);
-          callback(loadtemplates);
-        });
-      }
-      //bfestore.n32store(data, url, tempstore, callback);
-    },
-    error: function(XMLHttpRequest, textStatus, errorThrown) { 
-      bfelog.addMsg(new Error(), "ERROR", "FAILED to load external source: " + url);
-      bfelog.addMsg(new Error(), "ERROR", "Request status: " + textStatus + "; Error msg: " + errorThrown);
-    }
-  });
-}
-
-function retrieveLDS(uri, bfestore, loadtemplates, bfelog, callback){
-
-  var url = config.url + "/profile-edit/server/retrieveLDS";
-
-  $.ajax({
-    dataType: "json",
-    type: "GET",
-    async: false,
-    data: { uri: uri},
-    url: url,
-    success: function (data) {
-      bfelog.addMsg(new Error(), "INFO", "Fetched external source baseURI" + url);
-      bfelog.addMsg(new Error(), "DEBUG", "Source data", data);
-      bfestore.store = bfestore.jsonldcompacted2store(data, function(expanded) {
-        bfestore.store = [];
-        tempstore = bfestore.jsonld2store(expanded);
-        bfestore.storeDedup();
-        callback(loadtemplates);
-      });
-      //bfestore.n32store(data, url, tempstore, callback);
-    },
-    error: function(XMLHttpRequest, textStatus, errorThrown) {
-      bfelog.addMsg(new Error(), "ERROR", "FAILED to load external source: " + url);
-      bfelog.addMsg(new Error(), "ERROR", "Request status: " + textStatus + "; Error msg: " + errorThrown);
-    }
-  });
-}
-
-
-function deleteId(id, bfelog){
-  var url = config.url + "/verso/api/bfs/" + id;
-
-  $.ajax({
-    type: "DELETE",                
-    url: url,
-    dataType: "json",
-    success: function (data) {
-      bfelog.addMsg(new Error(), "INFO", "Deleted " + id);
-    },
-    error: function(XMLHttpRequest, textStatus, errorThrown) {
-      bfelog.addMsg(new Error(), "ERROR", "FAILED to delete: " + url);
-      bfelog.addMsg(new Error(), "ERROR", "Request status: " + textStatus + "; Error msg: " + errorThrown);
-    }
-  });
-
-}
-
 /* Config object profiles
  * Editor profiles are read from a WS endpoint
  * The data are expected to be in a JSON array, with each object
@@ -220,7 +36,7 @@ var config = {
               /* "logging": {
                 "level": "DEBUG",
                 "toConsole": false
-              }, */
+              },*/ 
   "url" : rectoBase,
   "baseURI": "http://id.loc.gov/",
   "basedbURI": "http://mlvlp04.loc.gov:8230",
@@ -404,35 +220,8 @@ var config = {
        }
 
      ]}
-
   ],
-  "save": {
-    "callback": save
-  },
-  "publish": {
-    "callback": publish
-  },
-  "retrieveLDS": {
-    "callback":retrieveLDS
-  },            
-  "retrieve": {
-    "callback": retrieve
-  },
-  "deleteId": {
-    "callback": deleteId
-  },            
-  /*            "load": [
-                {
-                "templateID": ["profile:bf:Work:Monograph", "profile:bf:Instance:Monograph", "profile:bf:Annotation:AdminMeta"],
-                "defaulturi": "http://id.loc.gov/resources/bibs/5226",
-                "_remark": "Source must be JSONLD expanded, so only jsonp and json are possible requestTypes",
-                "source": {
-                "location": "http://id.loc.gov/resources/bibs/5226.bibframe_raw.jsonp",
-                "requestType": "jsonp",
-                "data": "UNUSED, BUT REMEMBER IT"
-                }
-                }
-                ],*/
+  "api": ["save", "publish", "retrieveLDS", "retrieve", "deleteId"],
   "return": {
     "format": "jsonld-expanded",
     "callback": myCB
